Moving entities to a single file
