+++ .env.example
@@ -5,7 +5,10 @@ AWS_REGION=us-east-1
 AWS_ENDPOINT_URL=http://payment-localstack:4566
 
 #sns
-SNS_TOPIC=update_order_status
+UPDATE_ORDER_STATUS_TOPIC=arn:aws:sns:us-east-1:000000000000:update_order_status
+
+#sqs
+PRODUCTION_FAILED_QUEUE=http://sqs.us-east-1.localhost.localstack.cloud:4566/000000000000/production-failed-queue
 
 #orders
 ORDERS_URL=http://mock-server:8002
