+++ infra/docker/local/localstack.sh
@@ -2,7 +2,7 @@
 
 # -- > Create DynamoDb Table
 echo Creating  DynamoDb \'payments\' table ...
-echo $(aws --endpoint-url=http://localhost:4566 dynamodb create-table --cli-input-json '{"TableName":"payments", "KeySchema":[{"AttributeName":"order_id","KeyType":"HASH"}], "AttributeDefinitions":[ {"AttributeName":"order_id","AttributeType":"S"}], "BillingMode":"PAY_PER_REQUEST"}' --profile test-profile --region us-east-1 --output table | cat)
+echo $(aws --endpoint-url=http://localhost:4566 dynamodb create-table --cli-input-json '{"TableName":"payments", "KeySchema":[{"AttributeName":"order_id","KeyType":"HASH"}], "AttributeDefinitions":[ {"AttributeName":"order_id","AttributeType":"N"}], "BillingMode":"PAY_PER_REQUEST"}' --profile test-profile --region us-east-1 --output table | cat)
 
 # --> List DynamoDb Tables
 echo Listing tables ...
