--- internal/entities/entity/client.go
@@ -1,8 +1,0 @@
-package entity
-
-type Client struct {
-	ID    int    `json:"id"`
-	Cpf   int    `json:"cpf"`
-	Name  string `json:"name"`
-	Email string `json:"email"`
-}
