--- internal/entities/entity/client_test.go
@@ -1,17 +1,0 @@
-package entity
-
-import (
-	"github.com/stretchr/testify/assert"
-	"testing"
-)
-
-func TestClient(t *testing.T) {
-	c := Client{
-		ID:    1,
-		Cpf:   123456789,
-		Name:  "Person McPerson",
-		Email: "mcperson@example.com",
-	}
-
-	assert.IsType(t, c, Client{})
-}
